+++ src/main/code/index.ts
@@ -1,2 +1,6 @@
 export * from './code-types';
 export * from './code-utils';
+export {encodeAlpha} from './encodeAlpha';
+export {assembleJs} from './assembleJs';
+export {createVarRenamer} from './createVarRenamer';
+export {compileFunction} from './compileFunction';
