+++ src/test/code/compileFunction.test.ts
@@ -1,53 +1,4 @@
-import {assembleJs, compileFunction, createVar, createVarRenamer, encodeLowerAlpha} from '../../main/code';
-
-describe('encodeLowerAlpha', () => {
-
-  test('encodes numbers as latin letters', () => {
-    expect(encodeLowerAlpha(0)).toBe('a');
-    expect(encodeLowerAlpha(25)).toBe('z');
-
-    expect(encodeLowerAlpha(26)).toBe('aa');
-    expect(encodeLowerAlpha(51)).toBe('az');
-
-    expect(encodeLowerAlpha(52)).toBe('ba');
-    expect(encodeLowerAlpha(Number.MAX_SAFE_INTEGER)).toBe('bktxhsoghkkf');
-  });
-
-  test('encodes floating numbers', () => {
-    expect(encodeLowerAlpha(10.7)).toBe('k');
-  });
-});
-
-describe('createVarRenamer', () => {
-
-  test('return unique name', () => {
-    const varRenamer = createVarRenamer();
-    const var1 = createVar();
-    const var2 = createVar();
-
-    expect(varRenamer(var1)).toBe('a');
-    expect(varRenamer(var2)).toBe('b');
-    expect(varRenamer(var1)).toBe('a');
-    expect(varRenamer(var2)).toBe('b');
-  });
-});
-
-describe('assembleJs', () => {
-
-  test('assembles primitives', () => {
-    expect(assembleJs(true, createVarRenamer())).toBe('true');
-    expect(assembleJs(false, createVarRenamer())).toBe('false');
-    expect(assembleJs(123, createVarRenamer())).toBe('123');
-    expect(assembleJs('abc', createVarRenamer())).toBe('abc');
-  });
-
-  test('assembles variables', () => {
-    const aVar = createVar();
-    const bVar = createVar();
-
-    expect(assembleJs([aVar, '=', bVar, '+', aVar], createVarRenamer())).toBe('a=b+a');
-  });
-});
+import {compileFunction, createVar} from '../../main/code';
 
 describe('compileFunction', () => {
